test(sqs): move FIFO MD5 SDK tests to compatibility-tests/sdk-test-java
Address maintainer review on #786:
- Drop the software.amazon.awssdk:sqs test-scope dependency from the
  main module's pom.xml.
- Move SqsSdkMd5IntegrationTest to the compatibility-tests/sdk-test-java
  module as SqsMd5Test, where the existing AWS SDK v2 compatibility
  suite already lives. The test now follows that module's conventions
  (TestFixtures.sqsClient, AssertJ, lowerCamelCase method names, queue
  cleanup) and exercises the same scenarios: standard / FIFO / fair MD5
  parity, binary and custom-type-name attributes, and the dedup-replay
  case that asserts the second SendMessage returns the original
  MessageId and SequenceNumber.
